requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
drf-nested==1.2.22
drf-nested-routers==0.93.4
install==1.3.5
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.2.1
routers==0.10.1
sqlparse==0.4.3
tzdata==2022.4
